:art: Move make:model command to tempest/database package
